select widget with key / value does not show initial value #90
intermediate workaround
the fundamental issue is that we load the choices onFocus. This means
that initially, when the form is displayed, the UI does not know that
display name. Therefore, we default to name = value. onFocus, the
correct display name is presented.
This behavior is in line with an input autocomplete